Arregla las rutas que con el refactor se habian roto
